# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CoffeeFlow Backend API",
    "version": "1.0.0",
    "description": "Навчальний REST API. Ціни в копійках UAH; час з часовим поясом. Cookie-сесія та X-CSRFToken. Без JWT і без реєстрації клієнтів."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5000/api/v1",
      "description": "Локальна розробка"
    }
  ],
  "paths": {
...
    "/categories": {
      "get": {
        "summary": "Список категорій",
        "description": "",
        "security": [],
        "responses": {
          "200": {
            "description": "Успішно.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryList"
                },
                "example": {
                  "items": [
                    {
                      "id": 1,
                      "name": "Кава"
                    }
                  ],
                  "total": 1,
                  "limit": 20,
                  "offset": 0
                }
              }
            }
          },
          "422": {
            "description": "Помилка валідації.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Внутрішня помилка.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "База недоступна; Retry-After: 1.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 1000000,
              "default": 0
            }
          }
        ]
      },
      "post": {
        "summary": "Створити категорію",
        "description": "",
        "security": [
          {
            "cookieSession": [],
            "csrfHeader": []
          }
        ],
        "responses": {
          "201": {
            "description": "Успішно.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                },
                "example": {
                  "id": 1,
                  "name": "Кава"
                }
              }
            },
            "headers": {
              "Location": {
                "description": "URL створеного ресурсу.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "CSRF відсутній / неправильний або пошкоджений JSON.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Потрібен вхід або неправильні облікові дані.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Конфлікт даних, недоступний товар або заборонений перехід статусу.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Тіло перевищує 32 КіБ.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "Очікується Content-Type: application/json.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Помилка валідації.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Внутрішня помилка.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "База недоступна; Retry-After: 1.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryInput"
              },
              "example": {
                "name": "Чай"
              }
            }
          }
        }
      }
    },
    "/categories/{category_id}": {
      "get": {
        "summary": "Отримати категорію",
        "description": "",
        "security": [],
        "responses": {
          "200": {
            "description": "Успішно.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                },
                "example": {
                  "id": 1,
                  "name": "Кава"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не знайдено або замовлення належить іншому гостю.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Помилка валідації.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Внутрішня помилка.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "База недоступна; Retry-After: 1.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ]
      },
      "patch": {
        "summary": "Перейменувати категорію",
        "description": "",
        "security": [
          {
            "cookieSession": [],
            "csrfHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "Успішно.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                },
                "example": {
                  "id": 1,
                  "name": "Кава"
                }
              }
            }
          },
          "400": {
            "description": "CSRF відсутній / неправильний або пошкоджений JSON.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Потрібен вхід або неправильні облікові дані.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не знайдено або замовлення належить іншому гостю.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Конфлікт даних, недоступний товар або заборонений перехід статусу.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Тіло перевищує 32 КіБ.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "Очікується Content-Type: application/json.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Помилка валідації.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Внутрішня помилка.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "База недоступна; Retry-After: 1.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryInput"
              },
              "example": {
                "name": "Чай"
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Видалити порожню категорію",
        "description": "409, якщо існують товари, включно з архівними.",
        "security": [
          {
            "cookieSession": [],
            "csrfHeader": []
          }
        ],
        "responses": {
          "204": {
            "description": "Успішно; тіло відсутнє."
          },
          "400": {
            "description": "CSRF відсутній / неправильний або пошкоджений JSON.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Потрібен вхід або неправильні облікові дані.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не знайдено або замовлення належить іншому гостю.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Конфлікт даних, недоступний товар або заборонений перехід статусу.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Помилка валідації.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Внутрішня помилка.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "База недоступна; Retry-After: 1.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ]
      }
    },
    "/products": {
      "get": {
        "summary": "Список товарів",
        "description": "Архівні товари виключено; без available повертаються доступні й тимчасово недоступні.",
        "security": [],
        "responses": {
          "200": {
            "description": "Успішно.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductList"
                },
                "example": {
                  "items": [
                    {
                      "id": 1,
                      "category_id": 1,
                      "name": "Еспресо",
                      "description": "Класична чорна кава, 30 мл",
                      "price_cents": 4500,
                      "image_url": "",
                      "is_available": true
                    }
                  ],
                  "total": 1,
                  "limit": 20,
                  "offset": 0
                }
              }
            }
          },
          "422": {
            "description": "Помилка валідації.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Внутрішня помилка.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "База недоступна; Retry-After: 1.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 1000000,
              "default": 0
            }
          },
          {
            "name": "category_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 2147483647
            }
          },
          {
            "name": "available",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          }
        ]
      },
      "post": {
        "summary": "Створити товар",
        "description": "",
        "security": [
          {
            "cookieSession": [],
            "csrfHeader": []
          }
        ],
        "responses": {
          "201": {
            "description": "Успішно.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                },
                "example": {
                  "id": 1,
                  "category_id": 1,
                  "name": "Еспресо",
                  "description": "Класична чорна кава, 30 мл",
                  "price_cents": 4500,
                  "image_url": "",
                  "is_available": true
                }
              }
            },
            "headers": {
              "Location": {
                "description": "URL створеного ресурсу.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "CSRF відсутній / неправильний або пошкоджений JSON.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Потрібен вхід або неправильні облікові дані.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не знайдено або замовлення належить іншому гостю.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Конфлікт даних, недоступний товар або заборонений перехід статусу.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Тіло перевищує 32 КіБ.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "Очікується Content-Type: application/json.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Помилка валідації.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Внутрішня помилка.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "База недоступна; Retry-After: 1.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductInput"
              },
              "example": {
                "category_id": 1,
                "name": "Еспресо",
                "description": "Класична чорна кава, 30 мл",
                "price_cents": 4500,
                "image_url": "",
                "is_available": true
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "fields": {
                "type": "object",
                "additionalProperties": true
              }
            },
            "required": [
              "code",
              "message",
              "fields"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
      "Category": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "minimum": 1,
            "maximum": 2147483647
          },
          "name": {
            "type": "string",
            "maxLength": 80
          }
        },
        "required": [
          "id",
          "name"
        ],
        "additionalProperties": false
      },
      "CategoryInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          }
        },
        "required": [
          "name"
        ],
        "additionalProperties": false
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "minimum": 1,
            "maximum": 2147483647
          },
          "category_id": {
            "type": "integer",
            "minimum": 1,
            "maximum": 2147483647
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "default": ""
          },
          "price_cents": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10000000
          },
          "image_url": {
            "type": "string",
            "maxLength": 500,
            "description": "HTTPS URL без username або порожній рядок.",
            "default": ""
          },
          "is_available": {
            "type": "boolean",
            "default": true
          }
        },
        "required": [
          "id",
          "category_id",
          "name",
          "description",
          "price_cents",
          "image_url",
          "is_available"
        ],
        "additionalProperties": false
      },
      "ProductInput": {
        "type": "object",
        "properties": {
          "category_id": {
            "type": "integer",
            "minimum": 1,
            "maximum": 2147483647
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "default": ""
          },
          "price_cents": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10000000
          },
          "image_url": {
            "type": "string",
            "maxLength": 500,
            "description": "HTTPS URL без username або порожній рядок.",
            "default": ""
          },
          "is_available": {
            "type": "boolean",
            "default": true
          }
        },
        "required": [
          "category_id",
          "name",
          "price_cents"
        ],
        "additionalProperties": false
      },
...
      "CategoryList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Category"
            }
          },
          "total": {
            "type": "integer",
            "minimum": 0,
            "maximum": 2147483647
          },
          "limit": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          },
          "offset": {
            "type": "integer",
            "minimum": 0,
            "maximum": 1000000
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "additionalProperties": false
      },
      "ProductList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Product"
            }
          },
          "total": {
            "type": "integer",
            "minimum": 0,
            "maximum": 2147483647
          },
          "limit": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          },
          "offset": {
            "type": "integer",
            "minimum": 0,
            "maximum": 1000000
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "cookieSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "coffeeflow_session"
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRFToken"
      }
    }
  }
}